+++ semanticRoutine.c
@@ -9,7 +9,7 @@ static int debugSMR = 0;
 
 static MORPHEM morph;
 static int constArraySize;
-static long *constArray;
+static int32_t *constArray;
 static namelistProcedure *currentProcedure;
 static short codeLength;
 static labellist *labelList;
@@ -169,7 +169,7 @@ int blockAcceptProcedure() {
 		getNext(pList);
 	}while(getNext(pList) != NULL);
    
-   
+
 	// cause success should be 1 if everything worked good
 	return 1;
 }
@@ -471,7 +471,7 @@ int fa1(void){
 		// new item for constArray current length is on index 0
 		// update length of array in item 0
 		constArraySize++;
-		long* tmpArray = (long* )realloc(constArray, constArraySize * sizeof(long));
+		int32_t* tmpArray = (int32_t* )realloc(constArray, constArraySize * sizeof(int32_t));
 
 		if(tmpArray == NULL){
 			printf("\n\n\nfailure\n\n\n");
